\name{CommonJavaJars-package}
\alias{CommonJavaJars-package}
\docType{package}
\title{
Useful libraries for building a Java based GUI under R
}
\description{
Useful libraries for building a Java based GUI under R
}
\details{
\tabular{ll}{
Package: \tab CommonJavaJars\cr
Type: \tab Package\cr
License: \tab GPL-2\cr
}

Since the CommonJavaJars package has relatively big library files (~ 7MByte)
and should rarely be updated to save space in the CRAN archive.

}
\author{
Kornelius Rohmeyer (author only of two libraries and of the R code for loading them), see the copyright note for the authors of the other java libraries.

Maintainer: Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\note{
We provide the following Java libraries:
\itemize{
 \item Apache Commons Logging under the Apache License, Version 2.0, January 2004, \url{https://commons.apache.org/logging/}, Copyright 2001-2007 The Apache Software Foundation 
 \item Apache jog4j 2 under Apache License 2.0, \url{https://logging.apache.org/log4j/2.x/}, Copyright 1999-2005 The Apache Software Foundation
 \item Apache Commons Lang under  Apache License 2.0, \url{https://commons.apache.org/lang/}, Copyright 2001-2011 The Apache Software Foundation
 \item Apache Commons Collections under the Apache License, Version 2.0, January 2004, \url{https://commons.apache.org/collections/}, Copyright 2001-2008 The Apache Software Foundation
 \item Apache Commons Validator under the Apache License, Version 2.0, January 2004, \url{https://commons.apache.org/validator/}, Copyright 2001-2010 The Apache Software Foundation
 \item Apache Commons IO under the Apache License, Version 2.2 (last compatible with JDK 1.5), 2004, \url{https://commons.apache.org/io/}, Copyright 2002-2012 The Apache Software Foundation
 \item JLaTeXMath under GPL >= 2.0, \url{https://forge.scilab.org/index.php/p/jlatexmath/}, Copyright 2004-2007, 2009 Calixte, Coolsaet, Cleemput, Vermeulen and Universiteit Gent
 \item iText 2.1.4 under LGPL, \url{https://itextpdf.com/}, Copyright by Bruno Lowagie 
 \item SwingWorker under LGPL, formerly \verb{https://java.net/projects/swingworker/}, Copyright (c) 2005 Sun Microsystems 
 \item JXLayer under BSD License, formerly \verb{https://java.net/projects/jxlayer/}, Copyright 2006-2009, Alexander Potochkin 
 \item JGoodies Forms and Commons under BSD License, \url{http://www.jgoodies.com/freeware/libraries/forms/}, Copyright JGoodies Karsten Lentzsch  
 \item JSON Processing dual licensed under CDDL 1.1 and GPL 2 with Classpath Exception, \url{https://javaee.github.io/jsonp/}, Copyright (c) 2011-2013 Oracle
 \item AFCommons under the LGPL >= 2.1, \url{https://github.com/kornl/afcommons/}, Copyright (c) 2007-2014 by Kornelius Rohmeyer, Bernd Bischl, Santhosh Kumar T (ComponentTitledBorder class), Jerry Huxtable (filter code), Sebastian Haufe (JListDnD class), Romain Guy (Star class), Oracle (CloseTabPanel)
 \item JHLIR under BSD License, \url{https://jhlir.r-forge.r-project.org/}, Copyright (c) 2008-2010 by Bernd Bischl and Kornelius Rohmeyer  
 \item JRIEngine and REngine under LGPL-2.1, \url{https://www.rforge.net/rJava/}, Copyright 2010 by Simon Urbanek. Please use these two jar files only if package rJava does not provide them. 
 } 
}
\keyword{ package }

\examples{

\dontrun{
showGuideForJavaGUIs()

}
}
