\name{showGuideForJavaGUIs}
\alias{showGuideForJavaGUIs}

\title{
	Shows the Guide for Creating Java GUIS.
}
\description{
	Shows the guide for creating Java GUIs.
}
\usage{
	showGuideForJavaGUIs ()
}

\details{
	Opens the URL to the guide for creating Java GUIs.
	(To take less space in the archive of CRAN and its mirrors,
	this package should be updated rarely.
	Therefore the often updated guide itself is not directly part of the package.)
}
\value{
	The function itself returns NULL.
}
\author{
	Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}

\examples{

\dontrun{
showGuideForJavaGUIs() 

}
}
\keyword{ misc }


